# replace bool to empty struct in Hash and HashGeneric (optimization)