--- a/Readme.md
+++ b/Readme.md
@@ -34,16 +34,17 @@ Go v1.18 now supports generics which increases performance compared to reflectio
 
 The average time and allocated bytes are more than halved when using generics, and the number of times bytes have to be allocated from the heap are improved significantly.
 
-| **Name**                              	| **Runs** 	| **Average** 	| **Allocated** 	| **Allocations from heap** 	|
-|---------------------------------------	|-----------------:	|--------------------:	|---------------------:	|-------------------------------------------:	|
-| BenchmarkHash/Size_1-_interface-8     	| 4.631.326         	| 248.1 ns/op        	| 80 B/op             	| 5 allocs/op                               	|
-| BenchmarkHash/Size_1-_generics-8      	| 22.551.091        	| 46.19 ns/op        	| 8 B/op              	| 1 allocs/op                               	|
-| BenchmarkHash/Size_10-_interface-8    	| 581.262          	| 2.042 ns/op         	| 547 B/op            	| 24 allocs/op                              	|
-| BenchmarkHash/Size_10-_generics-8     	| 2.043.096         	| 549.6 ns/op        	| 187 B/op            	| 2 allocs/op                               	|
-| BenchmarkHash/Size_100-_interface-8   	| 59.907           	| 18.721 ns/op        	| 7.325 B/op           	| 213 allocs/op                             	|
-| BenchmarkHash/Size_100-_generics-8    	| 126.622          	| 8.221 ns/op         	| 3.359 B/op           	| 19 allocs/op                              	|
-| BenchmarkHash/Size_1000-_interface-8  	| 7.114            	| 22.1721 ns/op       	| 112.405 B/op         	| 2.038 allocs/op                            	|
-| BenchmarkHash/Size_1000-_generics-8   	| 13.345           	| 91.197 ns/op        	| 53.323 B/op          	| 74 allocs/op                              	|
-| BenchmarkHash/Size_10000-_interface-8 	| 523             	| 2.067.330 ns/op      	| 875.731 B/op         	| 20.173 allocs/op                           	|
-| BenchmarkHash/Size_10000-_generics-8  	| 1.264            	| 828.214 ns/op       	| 427.541 B/op         	| 320 allocs/op                             	|
+| **Name**                              	 | **Runs** 	 | **Average** 	 | **Allocated** 	 | **Allocations from heap** 	 |
+|-----------------------------------------|-----------:|--------------:|----------------:|----------------------------:|
+| BenchmarkHash/Size_1-_interface-16      |    5845340 |   202.9 ns/op |         80 B/op |                 5 allocs/op |
+| BenchmarkHash/Size_1-_generics-16       |   29534389 |   39.86 ns/op |          8 B/op |                 1 allocs/op |
+| BenchmarkHash/Size_10-_interface-16     |     800064 |    1438 ns/op |        515 B/op |                24 allocs/op |
+| BenchmarkHash/Size_10-_generics-16      |    2459796 |   482.6 ns/op |        170 B/op |                 2 allocs/op |
+| BenchmarkHash/Size_100-_interface-16    |      81973 |   14782 ns/op |       6907 B/op |               212 allocs/op |
+| BenchmarkHash/Size_100-_generics-16     |     184526 |    6467 ns/op |       3052 B/op |                19 allocs/op |
+| BenchmarkHash/Size_1000-_interface-16   |       7489 |  164909 ns/op |     101369 B/op |              2036 allocs/op |
+| BenchmarkHash/Size_1000-_generics-16    |      16734 |   70601 ns/op |      47743 B/op |                66 allocs/op |
+| BenchmarkHash/Size_10000-_interface-16  |        754 | 1592308 ns/op |     836264 B/op |             20216 allocs/op |
+| BenchmarkHash/Size_10000-_generics-16   |       1802 |  652946 ns/op |     389466 B/op |               255 allocs/op |
+